📝 구글차트 컨트롤 추가

+++ coin-chart@V0.2/coin-chart/src/main/java/com/example/coinchart/service/CoinService.java
@@ -2,12 +2,13 @@ package com.example.coinchart.service;
 
 import com.example.coinchart.domain.Coin;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
 public interface CoinService {
 
-    Map<String, Double> getExchanges();
+    Map<String, BigDecimal> getExchanges();
 
     List<Coin> getCoins();
